+++ packages/lumx-react/src/components/date-picker/DatePicker.tsx
@@ -29,6 +29,9 @@ interface DatePickerProps extends GenericProps {
     /** Value. */
     value: moment.Moment | undefined;
 
+    /** Month to display by default */
+    defaultMonth?: moment.Moment;
+
     /** On change. */
     onChange(value: moment.Moment | undefined): void;
 }
@@ -53,22 +56,26 @@ const DEFAULT_PROPS: Partial<DatePickerProps> = {
 
 /**
  * Simple component used to pick a date (semi-controlled implementation).
+ * @param props See DatePickerProps.
  *
  * @return The component.
  */
 const DatePicker = (props: DatePickerProps) => {
-    const today = props.value || moment();
+    const today = props.value || props.defaultMonth || moment();
     const [monthOffset, setMonthOffset] = useState(0);
 
     const setPrevMonth = () => setMonthOffset(monthOffset - 1);
     const setNextMonth = () => setMonthOffset(monthOffset + 1);
 
+    const selectedMonth = moment(today)
+        .locale(props.locale)
+        .add(monthOffset, 'months');
+
     return (
         <DatePickerControlled
-            monthOffset={monthOffset}
-            today={today}
             onPrevMonthChange={setPrevMonth}
             onNextMonthChange={setNextMonth}
+            selectedMonth={selectedMonth}
             {...props}
         />
     );